+++ src/components/LinkButton/index.ts
@@ -1,0 +1,2 @@
+export * from './LinkButton.types';
+export * from './LinkButton';