+++ src/octuple.ts
@@ -9,6 +9,7 @@ import {
   Avatar,
   AvatarGroup,
   AvatarGroupVariant,
+  AvatarPopupProps,
   StatusItemIconAlign,
   StatusItemsPosition,
 } from './components/Avatar';
@@ -16,10 +17,11 @@ import {
 import { Badge } from './components/Badge';
 
 import {
+  Button,
   ButtonShape,
   ButtonSize,
   ButtonTextAlign,
-  ButtonType,
+  ButtonVariant,
   ButtonWidth,
   ButtonIconAlign,
   DefaultButton,
@@ -79,6 +81,16 @@ import { Label, LabelSize } from './components/Label';
 import Layout from './components/Layout';
 
 import { Link } from './components/Link';
+
+import {
+  LinkButton,
+  LinkButtonIconAlign,
+  LinkButtonShape,
+  LinkButtonSize,
+  LinkButtonTextAlign,
+  LinkButtonVariant,
+  LinkButtonWidth,
+} from './components/LinkButton';
 
 import { List } from './components/List';
 
@@ -218,7 +230,7 @@ import { useBoolean } from './hooks/useBoolean';
 
 import { useCanvasDirection } from './hooks/useCanvasDirection';
 
-import { useMatchMedia } from './hooks/useMatchMedia';
+import { Breakpoints, useMatchMedia } from './hooks/useMatchMedia';
 
 import { useOnClickOutside } from './hooks/useOnClickOutside';
 
@@ -234,12 +246,16 @@ export {
   Avatar,
   AvatarGroup,
   AvatarGroupVariant,
+  AvatarPopupProps,
   Badge,
+  Breakpoints,
+  Button,
   ButtonIconAlign,
   ButtonShape,
   ButtonSize,
   ButtonTextAlign,
-  ButtonType,
+  ButtonVariant,
+  ButtonVariant as ButtonType, // TODO: Remove in Octuple v3.0.0, leave in for now to support legacy variant by <Tag /> implementations
   ButtonWidth,
   Card,
   CardSize,
@@ -284,6 +300,13 @@ export {
   LabelSize,
   Layout,
   Link,
+  LinkButton,
+  LinkButtonIconAlign,
+  LinkButtonShape,
+  LinkButtonSize,
+  LinkButtonTextAlign,
+  LinkButtonVariant,
+  LinkButtonWidth,
   List,
   Loader,
   LoaderSize,